+++ app/src/main/java/in/principal/dao/ActivityMarkDao.java
@@ -19,7 +19,7 @@ public class ActivityMarkDao {
         return avg;
     }
 
-    public static int getStudActMark(int studentId, long activityId, SQLiteDatabase sqliteDatabase){
+    public static int getStudActMark(long studentId, long activityId, SQLiteDatabase sqliteDatabase){
         int i = 0;
         Cursor c = sqliteDatabase.rawQuery("select Mark from activitymark where StudentId="+studentId+" and ActivityId="+activityId,null);
         c.moveToFirst();
@@ -31,7 +31,7 @@ public class ActivityMarkDao {
         return i;
     }
 
-    public static int getStudActAvg(int studentId, long activityId, SQLiteDatabase sqliteDatabase){
+    public static int getStudActAvg(long studentId, long activityId, SQLiteDatabase sqliteDatabase){
         int i = 0;
         Cursor c = sqliteDatabase.rawQuery("select (Avg(A.Mark)/B.MaximumMark)*100 as avg from activitymark A, activity B where A.ActivityId=B.ActivityId and A.ActivityId="+activityId+
                 " and StudentId="+studentId, null);
