+++ app/src/main/java/in/principal/fragment/VoiceSms.java
@@ -47,7 +47,7 @@ public class VoiceSms extends Fragment {
     private ArrayList<Section> secList;
     private ArrayList<Integer> classIdList;
     private ArrayList<Integer> secIdList;
-    private ArrayList<Integer> studIdList;
+    private ArrayList<Long> studIdList;
     private ArrayList<String> classNameList;
     private ArrayList<String> secNameList;
     private ArrayList<String> studNameList;
